pyproject.toml
[project]
name = "depenedency-heaven"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aci-sdk>=1.0.0b3",
    "fastapi>=0.116.1",
    "openai>=1.99.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
